pyproject.toml
[project]
name = "toolkit-hub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.1",
    "cryptography>=45.0.4",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "mysqlclient>=2.2.7",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-jose>=3.5.0",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.hatch.build.targets.wheel.sources]
src = "backend/src"
